+++ packages/create-deskapp/template/src/ui/src/app.ts
@@ -1,10 +1,12 @@
 import { createApp } from 'vue';
 import App from './App.vue';
-import faviconUrl from '../../icons/favicon.svg';
-import { pulse, PulsePlugin } from './plugins/pulse.ts';
+import faviconUrl from '@/ui/assets/images//favicon/favicon.ico';
+import { pulse, PulsePlugin } from '@/ui/plugins/pulse.ts';
 
+import "bootstrap/dist/css/bootstrap.min.css";
 import "bootstrap-icons/font/bootstrap-icons.css";
-import "./assets/scss/custom.scss";
+import "bootstrap/dist/js/bootstrap.bundle.min.js";
+import "@/ui/assets/scss/custom.scss";
 
 let favicon = document.querySelector<HTMLLinkElement>('link[rel="icon"]');
 if (!favicon) {